Reorganizing folder/package structure for better app structure
