+++ app/src/main/java/com/handigarde/todoapp/activities/EditItemActivity.java
@@ -1,4 +1,4 @@
-package com.handigarde.todoapp;
+package com.handigarde.todoapp.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,8 @@ import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.Spinner;
+
+import com.handigarde.todoapp.R;
 
 public class EditItemActivity extends AppCompatActivity {
