+++ app/src/main/res/layout/activity_edit_item.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.handigarde.todoapp.EditItemActivity">
+    tools:context="com.handigarde.todoapp.activities.EditItemActivity">
 
     <EditText
         android:layout_width="wrap_content"
